pyproject.toml
[tool.poetry]
name = "repl-nix-greenyellowfirstknowledge"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.40.8"
pinecone-client = "^5.0.1"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
flask = "^3.0.3"
nltk = "^3.9.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
